Refactoring and test refs
